+++ projects/hx-ui/src/lib/filters/filters.model.ts
@@ -24,7 +24,9 @@ export class FiltersModel implements IFiltersConfig {
 
    setSelectedOption(option?: IFilterOption) {
     if (option) {
-      this.selected.selected = false;
+      if (this.selected) {
+        this.selected.selected = false;
+      }
       this.selected = option;
       option.selected =  true;
     } else {