+++ projects/hx-ui/src/lib/selectize/selectize-item.interface.ts
@@ -1,4 +1,5 @@
 export interface ISelectizeItem {
   label: string;
   value: string;
+  disabled?: boolean;
 }