+++ src/app/app-routing.module.ts
@@ -20,6 +20,7 @@ import {TextInputComponent} from './core/text-input/text-input.component';
 import {OnlineStatusComponent} from './core/online-status/online-status.component';
 import {ToastrComponent} from './core/toastr/toastr.component';
 import { DateRangePickerComponent } from './core/date-range-picker/date-range-picker.component';
+import {DialogsComponent} from './core/dialogs/dialogs.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/install-guide', pathMatch: 'full' },
@@ -28,6 +29,7 @@ const routes: Routes = [
   { path: 'date-pickers',  component: DatepickersComponent },
   { path:  'date-range-pickers', component: DateRangePickerComponent},
   { path: 'dropdowns',  component: DropdownsComponent },
+  { path: 'dialogs',  component: DialogsComponent },
   { path: 'modals',  component: ModalsComponent },
   { path: 'tabs',  component: TabsComponent },
   { path: 'typeaheads',  component: TypeaheadsComponent },