+++ src/app/app.component.ts
@@ -51,8 +51,13 @@ export class AppComponent {
           link: '/dropdowns'
         },
         {
+          name: 'Dialogs',
+          link: '/dialogs'
+        },
+        {
           name: 'Modals',
-          link: '/modals'
+          link: '/modals',
+          deprecated: true
         },
         {
           name: 'Pagination',