+++ src/app/core/core.module.ts
@@ -25,6 +25,8 @@ import { TextInputComponent } from './text-input/text-input.component';
 import { OnlineStatusComponent } from './online-status/online-status.component';
 import { ToastrComponent } from './toastr/toastr.component';
 import { DateRangePickerComponent } from './date-range-picker/date-range-picker.component';
+import { DialogsComponent } from './dialogs/dialogs.component';
+import {CustomDialogComponent} from './dialogs/custom-dialog/custom-dialog.component';
 
 @NgModule({
   imports: [
@@ -55,14 +57,18 @@ import { DateRangePickerComponent } from './date-range-picker/date-range-picker.
       OnlineStatusComponent,
       ToastrComponent,
       DateRangePickerComponent,
-
+      DialogsComponent,
+      CustomDialogComponent
   ],
   providers: [
      TabularService,
      InMemoryDataService
   ],
   exports: [],
-  entryComponents: [CustomModalComponent]
+  entryComponents: [
+    CustomModalComponent,
+    CustomDialogComponent
+  ]
 })
 export class CoreModule {
   static forRoot(): ModuleWithProviders {